Added Layers!
Layers are very important part of an engine.
They needed for a very important thing: ...Layers! Yeah, I know.
Let's imagine some game. It has main UI and other things, like game world, e.t.c
And we will put UI on one layer(Overlay, actually, because we want it to be first) and other things on another layers.
Main structure of layers will be like that:

start();						end();
|							|
|---------------|--------------------|------------------|
|    Layers	| free space/nothing |	   Overlays	|
|---------------|--------------------|------------------|

Now, when we render things from the start(); to the end(); so overlays always will be on top (VERY IMPORTANT PART!!!)
And now, when we got event, we go from end(); to start(); so button on the top overlay will get event first, and other overlays/layers will not. Clearly a magic!
Logical, but magic. We can illustrate it like that:

	|-----------------------|
	| Layer that will not	|
	| get event because it	|
	| on layer lower ;(	|
	|			|
	| |-------------------| |
	| | Button, that we   | |
	| | clicked on, and it| |
	| | get event because | |
	| | located on top    | |
	| | 	layer :)      | |
	| |		      | |
	| |	*Click!*      | |
	| |-------------------| |
	|-----------------------|

So, for that purposes we have Layer class, that we must inherit in Sandbox and override some functions.
And now, viola! We got layer that can deal with events!

Also I fixed special build keys in premake5.lua ("/MD" thing), because, you know, if in MSVC we don't add it we will get assertion error. How I understood, it's something to deal with multi-threaded apps and heap memory allocation(maybe not.)