Added imgui!
Now this project has imgui inside it. Added code to handle it, created dummy functions to represent events that we need to dispatch.
It was, well, easy? The most important thing -
OpenGL folder in future will contain normal realization of functions that used now (For now it's just a copy of imgui files from `backends` folder)

Also added new event! KeyTypedEvent - representation of a event, when we click on button for text typing
And now we literally can write text in text fields! (When we get function callbacks for this, of course)