Added events to imgui!
Now we can interact with imgui! Yeah!
Funny thing - now we have layer dependency problem, that we need to fix later
Also we need to rewrite keys from dirty-dependent glfw system to our precious, that will be API independent.
(also i forgot to commit previous commit normally, so 1 or 2 lines of code actually must be in this commit, not in previous)